+++ baseutils/wstr_util.h
@@ -16,7 +16,7 @@ int     wstr_startswith(const WCHAR *str, const WCHAR *txt);
 int     wstr_startswithi(const WCHAR *str, const WCHAR *txt);
 int     wstr_endswithi(const WCHAR *str, const WCHAR *end);
 int     wstr_empty(const WCHAR *str);
-int     wstr_copy(WCHAR *dst, int dst_cch_size, const WCHAR *src);
+int     wstr_copy(WCHAR *dst, size_t dst_cch_size, const WCHAR *src);
 int     wstr_copyn(WCHAR *dst, size_t dst_cch_size, const WCHAR *src, size_t src_cch_size);
 int     wstr_skip(const WCHAR **strp, const WCHAR *expect);
 int     wstr_copy_skip_until(const WCHAR **strp, WCHAR *dst, size_t dst_size, WCHAR stop);
